+++ changelog.md
@@ -5,6 +5,7 @@
 ## Upcoming
 * Curses!
   * Items with curses only can now be enchanted and get a 40% boost to max cost
+  * Added Curse of Mending. Consumes experience when used
 * Mining charge crafting is now cheaper
 * All items in creative tabs are now sorted!
 * Fixed beeg ore veins generating less at deepslate level